+++ Domibus-MSH/src/main/java/eu/domibus/core/party/ProcessRo.java
@@ -6,15 +6,15 @@ package eu.domibus.core.party;
  */
 public class ProcessRo {
 
-    private int entityId;
+    private long entityId;
 
     private String name;
 
-    public int getEntityId() {
+    public long getEntityId() {
         return entityId;
     }
 
-    public void setEntityId(int entityId) {
+    public void setEntityId(long entityId) {
         this.entityId = entityId;
     }
 
@@ -39,7 +39,7 @@ public class ProcessRo {
 
     @Override
     public int hashCode() {
-        int result = entityId;
+        int result = (int)entityId;
         result = 31 * result + (name != null ? name.hashCode() : 0);
         return result;
     }
